Revert "Merge pull request #3879 from AmitJoki/WebpackAnalysis"
This reverts commit a98131e43daf7ca6c9530284af26a87a95e62411, reversing
changes made to a40b8890e86937d0567a62cb1bce0b6898bdfbf2.
