+++ .gitignore
@@ -51,4 +51,3 @@ Session.vim
 1
 .byebug_history
 .ruby-gemset
-stats.json
